Use full path to execute file
What this does is use readlink -f "$0" to get the absolute path of the astra script. The dirname command is then used to extract the directory name from the absolute path, which is stored in the DIR variable. Finally, the ruby command is used to run the astra.rb file using the $DIR variable to reference the absolute path of the astra.rb file.

With these changes, the astra script should be able to find the astra.rb file regardless of the current working directory.